tp-5 caja mercado central parte 1 terminada.

+++ src/ar/edu/unq/po2/tp5/Producto.java
@@ -3,15 +3,15 @@ package ar.edu.unq.po2.tp5;
 public abstract class Producto {
 
 	protected String nombre;
-	protected Float precio;
+	protected Double precio;
 	
 
-	public Producto(String nombre, Float precio) {
+	public Producto(String nombre, Double precio) {
 		this.nombre = nombre;
 		this.precio = precio;
 
 	}
 	
-	public abstract Float getPrecio();
+	public abstract Double getPrecio();
 
 }
