+++ src/ar/edu/unq/po2/tp5/ProductoEmpresa.java
@@ -2,13 +2,13 @@ package ar.edu.unq.po2.tp5;
 
 public class ProductoEmpresa extends Producto {
 
-	public ProductoEmpresa(String nombre, Float precio) {
+	public ProductoEmpresa(String nombre, double precio) {
 		super(nombre, precio);
 		
 	}
 
 	@Override
-	public Float getPrecio() {
+	public Double getPrecio() {
 		return this.precio;
 	}
